tshirt_sales/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-experimental
langchain-huggingface
langchain-core
chromadb
sentence-transformers
sqlalchemy
mysql-connector-python
python-dotenv
notebook